requirements.txt
# Core
Django>=5.0,<5.1
wagtail>=6.0,<6.1

# Dev
factory-boy>=3.3,<3.4
Faker>=23.3,<23.4

# Changelog Generation
bumpver==2023.1129
invoke==2.2.0